requirements.txt
streamlit
nltk
textblob
sumy
scikit-learn
matplotlib
wordcloud
pandas
PyPDF2
python-docx
openpyxl
